DISCORD/OFICIAL/v1.4.1/requirements.txt
discord~=1.0.1
pillow~=7.2.0
pandas~=1.1.2
numpy~=1.19.2